pyproject.toml
[project]
name = "rag-tutorials-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.7",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.28",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "typesense>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pinecone>=5.0.0",
    "boto3>=1.35.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.9.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}
packages = ["rag_backend", "src"]
